Add files via upload
a python script that retrieves Fwds Curve and Spot single point data from the NCFX curve API, and then using the standard tenors, calculates the Straight line values between tenors. We convert the fwd points to true decimal and add to the spot rate to construct forward outright using both methods. Finally, we compare the outright values based on a notional trading amount of 1 million base.
We will need to adjust this so that the notional amounts are all in the same currency (USD)